requirements.txt
Flask
gunicorn
alpaca-trade-api
python-dotenv
